Add script for creating external network for jumbo MTU job
This is script that will be used temporary on a setup which is configured
to use jumbo MTU for creating an external network exactly like cloud-config
creates. It will be run as --overcloud-postdeploy-action.

Note: cloud-config can not be used due to a couple of reasons:
a. cloud-config on a setup with jumbo MTU creates an external network with
   mtu 9000 which is not what we need for our testing.
b. while it is still possible to add support to cloud-init to support
   setting mtu the feature depends on ansible 2.9 but infrared currently
   is hard-coded to use 2.7.* (probably due to compatibility reasons)
There is a patch[0] for allowing usage of latest ansible in infrared
but it is still in review (long time already).
[0] https://review.gerrithub.io/c/redhat-openstack/infrared/+/493623

Change-Id: I2825a99f60dc440b760f4e694eb857b3850bfa60