Update nfv tempest config file
Several changes have been made to the nfv-tempest-plugin.
Some tests have been removed as not reflecting the nfv testing
scenarios. And some configuration removed from the external config file.
Updating the configuration file according to the changes.

- Remove all numaX flavors.
  The numaX tests have been completely removed from the plugin as they
  are not reflecting the nfv testing scenarios anymore.
- A base "nfv_qe_base_flavor" flavor has been added.
  Will use as a baseline for most of the nfv tests.
- Tests specific configuration moving into the tempest.conf and as a
  result are moving out of the tempest_config file.
- Tests execution selection changed to the regex based string.

Note - As per now, neutron job executed the following tests:
- test_numa0_provider_network
- test_numamix_provider_network
- test_packages_compute
- test_cold_migration

With the latest changes of the nfv plugin, only the test_cold_migration
test is able to be executed on neutron based setup.

In case neutron setups will be tuned, more tests will be available for
the execution.

Change-Id: I41de062a5a28490ed332d49ab8b8455e3b36f0c5