ovn: Don't include os-net-config mapping because of SR-IOV
When migrating with SR-IOV, the PFs could be in use and not seen by the
host kernel. This causes the os-net-config to fail when the mapping is
provided during overcloud deploy step.

This patch removes the os-net-config mapping from the deploy command to
not run again configuration over SR-IOV interfaces that had already been
configured when deployed.

Change-Id: I4e57cc668b80d020ca163d8897020982655996ee